+++ Tests/SwiftSDKTests/RTTests/RTMessagingTests.swift
@@ -24,7 +24,7 @@ import XCTest
 
 class RTMessagingTests: XCTestCase {
     
-    private let backendless = Backendless.shared
+    /*private let backendless = Backendless.shared
     private let timeout: Double = 10.0
     private let CHANNEL_NAME = "TestsChannel"
     
@@ -171,5 +171,5 @@ class RTMessagingTests: XCTestCase {
             })
         })
         waitForExpectations(timeout: timeout, handler: nil)
-    }
+    }*/
 }