+++ Tests/SwiftSDKTests/UOWTests/DataUOWCreateTests.swift
@@ -24,7 +24,7 @@ import XCTest
 
 class DataUOWCreateTests: XCTestCase {
     
-    private let backendless = Backendless.shared
+    /*private let backendless = Backendless.shared
     private let testObjectsUtils = TestObjectsUtils.shared
     private let timeout: Double = 20.0    
     private let tableName = "TestClass"
@@ -309,5 +309,5 @@ class DataUOWCreateTests: XCTestCase {
             XCTFail("\(fault.code): \(fault.message!)")
         })
         waitForExpectations(timeout: timeout, handler: nil)
-    }
+    }*/
 }