+++ Tests/SwiftSDKTests/UOWTests/DataUOWUpdateTests.swift
@@ -24,7 +24,7 @@ import XCTest
 
 class DataUOWUpdateTests: XCTestCase {
     
-    private let backendless = Backendless.shared
+    /*private let backendless = Backendless.shared
     private let testObjectsUtils = TestObjectsUtils.shared
     private let timeout: Double = 20.0    
     private let tableName = "TestClass"
@@ -397,5 +397,5 @@ class DataUOWUpdateTests: XCTestCase {
             XCTFail("\(fault.code): \(fault.message!)")
         })
         waitForExpectations(timeout: timeout, handler: nil)
-    }
+    }*/
 }